requirements.txt
kafka-python==2.0.2
pandas
scikit-learn
fastapi
uvicorn
websockets
